refactor: rename CTA
